add user tabel, and update task management to associate tasks with users  now every user has their separate todo and users and their tasks are stored in database

+++ app/routes/tasks.py
@@ -6,19 +6,19 @@ task_bp = Blueprint('tasks',__name__)
 
 @task_bp.route('/')
 def view_tasks():
-    if 'user' not in session:
+    if 'user_id' not in session:
         return redirect(url_for('auth.login'))
     
-    tasks = Task.query.all()
+    tasks = Task.query.filter_by(user_id = session['user_id']).all()
     return render_template('tasks.html',tasks = tasks)
 @task_bp.route('/add',methods = ['POST'])
 def add_task():
-    if 'user' not in session:
+    if 'user_id' not in session:
         return redirect(url_for('auth.login'))
     
     title = request.form.get('title')
     if title:
-        new_task = Task(title = title,status = 'Pending')
+        new_task = Task(title = title,status = 'Pending',user_id = session['user_id'])
         db.session.add(new_task)
         # It registers this object as something you want to insert later
         db.session.commit()
@@ -27,7 +27,9 @@ def add_task():
 
 @task_bp.route('/toggle/<int:task_id>',methods = ['POST'])
 def toggle_status(task_id):
-    task = Task.query.get(task_id)
+    if 'user_id' not in session:
+        return redirect(url_for('auth.login'))
+    task = Task.query.filter_by(id = task_id,user_id = session['user_id']).first()
     if task:
         if task.status == 'Pending':
             task.status = 'Working'
@@ -42,7 +44,9 @@ def toggle_status(task_id):
         
 @task_bp.route('/clear',methods = ["POST"])
 def clear_tasks():
-    Task.query.delete()
+    if 'user_id' not in session:
+        return redirect(url_for('auth.login'))
+    Task.query.filter_by(user_id=session['user_id']).delete()
     db.session.commit()
     flash('All tasks cleared')
     return redirect(url_for('tasks.view_tasks'))
